+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Fix building with Bazel on Linux by excluding Xcode support.
 - Fix slow baseline filtering for large projects with many results.
 - Fix inconsistent unused parameter results when the function is declared in a file that is a member of multiple targets.
+- Fix inconsistent results caused by identical extensions declared in different modules.
 
 ## 3.1.0 (2025-04-05)